[#28] refactor : account dao flow 데이터 추가
(cherry picked from commit 8051d204bbd0e0d178d9d8f11c493b35475278fb)
