+++ app/src/main/java/com/euphony/project/account_touch/data/account/entity/Account.kt
@@ -1,11 +1,13 @@
-package com.euphony.project.account_touch.data.entity
+package com.euphony.project.account_touch.data.account.entity
 
 import androidx.annotation.NonNull
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
-import com.euphony.project.account_touch.data.entity.model.Color
+import com.euphony.project.account_touch.data.bank.entity.Bank
+import com.euphony.project.account_touch.data.global.base.entity.BaseEntity
+import com.euphony.project.account_touch.utils.model.Color
 import java.util.*
 
 @Entity(
@@ -13,7 +15,7 @@ import java.util.*
     foreignKeys = [ForeignKey(
         entity = Bank::class,
         parentColumns = arrayOf("bank_id"),
-        childColumns = arrayOf("bank"),
+        childColumns = arrayOf("bank_id"),
         onDelete = ForeignKey.CASCADE
     )]
 )
@@ -24,17 +26,14 @@ data class Account(
     val id: Long = 0L,
 
     @NonNull
-    @ColumnInfo(name = "bank", index = true)
-    val bank: Long,
+    @ColumnInfo(name = "bank_id", index = true)
+    val bank_id: Long,
 
     @ColumnInfo(name = "nickname")
     val nickname: String,
 
     @ColumnInfo(name = "account_number")
     val accountNumber: String,
-
-    @ColumnInfo(name = "is_allow_any")
-    val isAllowAny: Boolean,
 
     @ColumnInfo(name = "is_always_on")
     val isAlwaysOn: Boolean,
@@ -48,4 +47,4 @@ data class Account(
     @ColumnInfo(name = "modify_date")
     override val modifyDate: Date = Date(System.currentTimeMillis()),
 
-) : BaseEntity()
+    ) : BaseEntity()
